feat(plugin): adapter plugin -- h2-tcp and clickhouse-http
