+++ configure/etc/conf/plugins.properties
@@ -43,6 +43,8 @@ datacap-plugin-starrocks=plugin/datacap-plugin-starrocks/pom.xml
 datacap-plugin-hologres=plugin/datacap-plugin-hologres/pom.xml
 datacap-plugin-greptimedb=plugin/datacap-plugin-greptimedb/pom.xml
 datacap-plugin-questdb=plugin/datacap-plugin-questdb/pom.xml
+datacap-plugin-h2-tcp=plugin/datacap-plugin-h2-tcp/pom.xml
+datacap-plugin-clickhouse-http=plugin/datacap-plugin-clickhouse-http/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
