+++ plugin/datacap-plugin-clickhouse-http/src/main/java/io/edurt/datacap/plugin/ClickHouseHTTPPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class ClickHouseHTTPPlugin
+        extends Plugin
+{}
