+++ plugin/datacap-plugin-clickhouse-http/src/main/resources/META-INF/services/io.edurt.datacap.plugin.Plugin
@@ -1,0 +1,1 @@
+io.edurt.datacap.plugin.ClickHouseHTTPPlugin
