+++ plugin/datacap-plugin-h2-tcp/src/main/java/io/edurt/datacap/plugin/H2TCPPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class H2TCPPlugin
+        extends Plugin
+{}
