+++ pom.xml
@@ -71,9 +71,9 @@
         <module>plugin/datacap-plugin-hologres</module>
         <module>plugin/datacap-plugin-greptimedb</module>
         <module>plugin/datacap-plugin-questdb</module>
-        <!--                <module>plugin/datacap-http-clickhouse</module>-->
+        <module>plugin/datacap-plugin-h2-tcp</module>
+        <module>plugin/datacap-plugin-clickhouse-http</module>
         <!--        <module>plugin/datacap-http-ceresdb</module>-->
-        <!--        <module>plugin/datacap-jdbc-h2</module>-->
         <!--        <module>plugin/datacap-jdbc-redis</module>-->
         <!--        <module>plugin/datacap-jdbc-mongo</module>-->
         <!--        <module>plugin/datacap-jdbc-elasticsearch</module>-->
